pyproject.toml
[project]
name = "vetspro"
version = "0.1.0"
description = ""
authors = [
    {name = "AsheshShrestha007",email = "ashesh.shrestha@qsystems.ai"}
]

readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "bcrypt (>=4.2.1,<5.0.0)",
    "pyodbc (>=5.2.0,<6.0.0)",
    "mysql-connector-python (>=9.2.0,<10.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "python-jose (>=3.3.0,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
